moved amqp to util in support test
