+++ test/support/util.ts
@@ -1,4 +1,5 @@
 import { AssertionError, expect } from "chai"
+import * as ampq from "amqplib"
 
 export function elapsedFrom(from: number): number {
   return Date.now() - from
@@ -50,4 +51,45 @@ export function range(count: number): number[] {
     ret[index] = index
   }
   return ret
+}
+
+export async function getMessageFrom(stream: string): Promise<{ content: string; properties: ampq.MessageProperties }> {
+  return new Promise(async (res, rej) => {
+    const con = await ampq.connect("amqp://rabbit:rabbit@localhost")
+    con.on("error", async (err) => rej(err))
+    const ch = await con.createChannel()
+    await ch.prefetch(1)
+    await ch.consume(
+      stream,
+      async (msg) => {
+        if (!msg) return
+        msg.properties.userId
+        ch.ack(msg)
+        await ch.close()
+        await con.close()
+        res({ content: msg.content.toString(), properties: msg.properties })
+      },
+      { arguments: { "x-stream-offset": "first" } }
+    )
+  })
+}
+
+export async function createClassicConsumer(
+  stream: string,
+  cb: (msg: ampq.Message) => void
+): Promise<{ conn: ampq.Connection; ch: ampq.Channel }> {
+  const conn = await ampq.connect("amqp://rabbit:rabbit@localhost")
+  const ch = await conn.createChannel()
+  await ch.prefetch(1)
+  await ch.consume(
+    stream,
+    async (msg) => {
+      if (!msg) return
+      cb(msg)
+      ch.ack(msg)
+    },
+    { arguments: { "x-stream-offset": "first" } }
+  )
+
+  return { conn, ch }
 }
